Split peer tests into submodules
